chore: javadsl, api docs, refactoring of ConsumerFilter (#823)
